Project Informations
"Projekt City" is an unfunded, free and massive Minecraft™ project! Our mission is to entertain as many players as possible and create a fun and easy to understand map that is completely free to play. 
The entire map is in English² and will remain so from now on, as it is too complicated to add other languages. However, all versions except V4.2 are currently in German as we forgot to update it. We hope you understand. Version V4.2 contains (so far) everything in English. 

We try to fulfil every wish as well and as quickly as possible. Requests can be made on our Discord³ server, on our website and in the Curseforge Comments. 
We are looking forward to your feedback, suggestions and wishes for this project, so that all players have fun on our map. We are trying to become more professional with every update and are constantly improving our building skills.